Migrate Pep9 highlighters to Pep9 library.
Code highlighting is highly version dependent.
However, it is a feature that ever Pep version supports.
For this reason, APepVersion gains two methods that create ASM/microcode highlighters.
These highlighters are able to be re-styled like current highlighters.
This allows the implementation details to be moved to the Pep/9 library.