refactor!: forbid multi-section programs
In an effort to reduce complexity, each source file is only allowed to have 1 section.
That section is assigned automatically by the assembler.
Giving students the ability to create multi-section programs without the ability to write linker scripts was a recipie for disaster.
Really, just giving students multiple secrtions was the bad call.

This simplifies the assembler because there is now a 1:1 mapping between images and top level sections.