build(ui): change package type to  module
Downstream projects crash when importing converters.
Downstream projects `import` this module in their sources.
This causes the *.es.js variant to be selected, which itself contains imports.

Prior to this commit, this packages was CommonJS.
This means that the *.js are interpreted as CommonJS.

Therefore, our *.es.js variant (with imports) is treated as CommonJS.
Additionally, main+module package fields seems to be messing with exports.

To resolve this we must make the package a module.
We must also fix our jest config to be module compliant.