refactor(asm.pas)!: switch to N-way tree for AST
There was a code explosion because of all the concrete base classes.
The mix of compile time and run time inheritance was making implementing ops on the trees hard.

Now, every node is of the same nominative type
Nodes have dynamic attributes.
The "real type" of a node is now determined by its set of attributes.
This may lead to a situation where multiple nodes share the same structural type.

Operations should be written such that they apply to anything with the matching structural type.
An operation does this by querying the types of the node's attributes.
In the case where exact matches are required (e.g., only EXPORT), the values of the node's attributes can be checked.

This implementation is inspired by the nominal type system from typescript.


BREAKING CHANGE: Entire tree API has been refactored.