feat(asm.pas): add visitor to bind & detect globals
If a symbol declare a global, update the hierarchy to which it belongs and intern in global symbol table.

If an instruction uses a symbolic argument, attempt to bind to global symbols.
Perform this binding even if the symbol is defined for ease of implementation.
In this case, mark an error about multiple definitions.

Progress on #12, #14.