refactor(asm.sym): add pointer size to symbols
There are now two different kinds of size: pointer and pointed.
pointed size is the size of the thing being represented.
i.e., 2 for a constant, 3 for a non-unary instruction.
pointer size is how many bytes to store the address to access pointed, which is always 2 on pep.

I need this delineation so that charIn/charOut report a size of 1 rather than 2.
This is needed because I construct my memory maps from symbol tables, and memory maps are wrong sized without this.