refactor(sim)!: allow long reads+writes for simulator
Leaving our length at quint8 prevents me from copying the OS in all at once.
Now use a Address, so we could (in theory) read/write the entire device in one op.

We'll have to throw some error if people try to do this with tracing on, because we can't store traces this large.

BREAKING CHANGE: length parameter on read+write is changed.