ci: use `vcpkg`to build boost locally
This should reduce the burden on other developers to get started with the project.

However, vcpkg downloads boost one submodule at a time.
This takes several minutes.
This would negatively impact our CI speed goals.

I have include a USE_SYSTEM_BOOST flag that is recognized by our buildsystem.
It may be defined as an environment variable, or passed to cmake with -D.
If it is defined, then we do not install boost via VCPKG.
You must ensure that Boost_ROOT is set correctly on your system.