ci: remove node; include Qt and build tools
I reduced the size of the base image by 70MB (2x) by switching to a base ubuntu image.
Additionally, I cut LLVM as we have not been using it.

However, I've included Qt.
This may increase the size of the final image, but I suspect this will be faster than installing Qt in each action.

I have used the tool `dive` to optimize the size of my image.
Hopefully this can reduce the pull time for actions.