feat(ui): add generic syntax highlighter based on Pep/9
This highlighter specifies rules in terms of a syntax type.
i.e., Symbol, Comment.

A style manager stores the current text format for each syntax type.
The manager can be updated independently of the  highlighter.
Hopefully we can tie this style manager into whatever we use for app coloring.

TODO: still need to implement the FSM aspect of my highlighting rules.
The FSM will allow me to handle states such as being in a block comment.