chore(ui): avoid backtracking in highlighting
Avoid re-matching already-highlighted text if possible.
This works out-of-the-box for most of the syntax.
however, we should allow the index to be reset selectively.
For example, a symbol in a comment will be highlighted incorrectly with reset.

In reality, I need an even more complex FSM of rules, looking like the parsing FSM.
However, this should somewhat speed parsing for now.