Treat all words as opcodes, not just negative ones
All negative opcodes are expected to execute natively.
Non-negative opcodes are treated as a CALL ADDR, where ADDR is the opcode.
This eliminates DOCOL, and eliminates inconsistencies between COREINT and other words transferring control.

This is a realistic change, since my Pep/10 version will be subroutine threaded.
This means a bunch of CALLs, which will handle modifying the instruction pointer.
DOCOL will also be a noop there.