Rearrange dictionary entry
Move name before the link pointer. This way it is simply offset math to find the CWA of a pointer.
We also insert the padding bytes before the string rather than within an entry.
However, we still respect any level of alignment requested for the code fields.

This new approach makes it easy to add additional fixed-length fields in the middle of an entry.
One such example would be a vtable pointer.

Implementation derived from GFORTH.
See:
  - The new Gforth Header; Anton Ertl
  - https://www.semanticscholar.org/paper/The-new-Gforth-Header-Ertl/c12d67bb1bba0e36350644b7d6dae2659919e931