chore(bin): flatten parse tree
Previously, all lines inside a macro were parented within a dummy "macro chunk".
This commit removes the macro chunk from the grammar.
All statements inside and outside a MACRO declaration will be at the same depth in the tree.

To put all "children" of the macro chunk inside a node in the AST will require manipulation of the parse tree.
Writing custom visitors requires I write more code than adding complexity to the grammar.
Debugging the grammar is difficult, while debugging visitors is not.
In my mind, this justifies the (potential) extra code.