refactor(sim): create templatized trace iterator
This iterator API is the result of multiple design iterations.

The old API returned a polymorphic iterator at the frame and packet level.
This creates many short-lived items on the heap.

A primary design goal was to avoid dynamic memory allocation for each iterator.
There will be thousands of frames with dozens of packets processed for each simulation.
It seems unlikely that we could meet performance expectations given that amount of dynamic allocation.
I could get around this with a custom (slab) allocator, but I'd rather design an API that does not need dynamic allocation.

I attempted to design a PIMPL version, but this had the same dynamic allocation issues.
An alternative implementation (patch below) created a new iterator class for:
 - frames
 - packets
 - payloads
This is a significant amount of code repetition.

I've realized that some object in the system must be polymorphic, since in-memory layout of a trace depends on the buffer used.
Seizing on this, I've moved polymorphic behavior into the buffer class (via IteratorImpl).
Buffer implementations should also inherit from IteratorImpl.
The buffer class is likely to be allocated dynamically, or at least outlive the iterators.
This allows me to move the polymorphic dispatch into the buffer from the iterator.

The current solution leverages structural similarities between the iterators.
Correct methods and implementations are selected via a "Level" tag.

```
diff --git a/logic/sim/api2.hpp b/logic/sim/api2.hpp
--- a/logic/sim/api2.hpp
+++ b/logic/sim/api2.hpp
@@ -207,4 +207,50 @@
     virtual bool analyze(const packet::Header&, const std::span<packet::Payload>&, Direction) = 0;
 };
+/*
+struct Packet {
+    Packet(IteratorImpl* impl, std::size_t location);
+    packet::Header header() const;
+    int payloadCount() const;
+    packet::Payload payload(int index) const;
+private:
+    const IteratorImpl* _impl;
+    const std::size_t _location = 0;
+};
+
+struct Frame {
+    typedef std::bidirectional_iterator_tag iterator_category;
+    typedef const Packet value_type;
+    typedef quint64 difference_type;
+    typedef const Packet* pointer;
+    typedef const Packet& reference;
+    Frame(IteratorImpl* impl, std::size_t location);
+    Frame& operator++();
+    Frame& operator++(int);
+    bool operator==(Frame other) const;
+    bool operator!=(Frame other) const;
+    value_type operator*() const;
+private:
+    const IteratorImpl* _impl;
+    const std::size_t _location = 0;
+};
+
+// Pass std::functions to iterator to do reads/writes
+struct Iterator {
+    typedef std::bidirectional_iterator_tag iterator_category;
+    typedef const Frame value_type;
+    typedef quint64 difference_type;
+    typedef const Frame* pointer;
+    typedef const Frame& reference;
+    Iterator(IteratorImpl* impl, std::size_t location);
+    Iterator& operator++();
+    Iterator& operator++(int);
+    bool operator==(Iterator other) const;
+    bool operator!=(Iterator other) const;
+    value_type operator*() const;
+
+private:
+    const IteratorImpl* _impl;
+    const std::size_t _location = 0;
+};*/

 // TODO: Add additional channel for command / simulation packets.
```