refactor(sim): unify filter/analyze in trace sink
These methods were combined because I can't think of a usecase for filter by itself.
It doesn't make sense to ask "can you process this packet" without wanting to do something to that packet later.

I've combined the two functions, as discussed in #406.
Now, instances should return false if they can't meaningfully handle the packet, or if they chose not to.
If the instance processes the packet---even if that processing is a no-op---it should return true.