refactor(ui): combine breakpoint and line number components
This reduces the number of spawned delegates by a factor of two.
It also reduces the amount of centering logic by a factor of two.
In release mode, the Pep10 OS loads *almost* instantly, with some lag due to highlighting.

I learned that heights & widths should be reals.
Otherwise, alignment between items will be wrong due to quantization errors.

When preferences are merged, I will need to connect the colors & styles.