feat(oc): add placeholder for object code editor
Refactored main screen mode switching to avoid nested ternaries.

Currently implemented via a text editor with r & rw modes.

Note that the text editor is nested inside an Item, not directly in a ScrollView.
When unnested, the ScrollView sets its height/width to the content width.
This occurs even when it has been *explicitly instructed* to fill the parent layout.

Progress on #393.