feat(oc): add cursor movement to editor
There's still some issues with the tableview.
On adding a new row, cells throughout the entire table are randomly updated.
The edit text is correct, while the display text isn't
Looking at logs, it appears that during the index renumbering, the model fails to do math correctly.

This bug also blocks release of the Memory Dump.
In that component, it shows up as randomization of text values on scroll.