refactor(oc): remove edit delegate due to ghosting
In a TableView, a `delegate` is used to render a cell.
A delegate allows an attached property called `editDelegate`.
Usually, an instance of the edit delegate is used to mutate the model.

In some cases edit delegates break delegate rendering.
Scrolling with edit delegates attached causes pooled delegates to be reused without updating delegate state.

The solution to this is to make the delegate an editable control, bypassing the edit delegate.