feat(oc): attempt to fix table delegate confusion
This is ultimately a failed attempt to prevent the delegate from "smearing".

TableView.editDelegate was introduced in Qt 6.5.0.
Since then, it has been used 74 times on public github projects
~50% of references are directly to Qt source code.
Of the remaining ~30, we make up 2 of those references.
Our editDelegate seems to be one of the most complex from an event perspective.
It will be difficult to fix the buggy behavior with an edit delegate.
This code is not "useful" in the sense that it still exhibits the bug.
However, it documents the steps I took attempting to work around it.

Most extant editable table views use a normal delegate that happens to be editable.
This was require before Qt 6.5, and seems more robust than edit delegates.
With an attached edit delegate, the index/row/column of a delegate may not be updated when it is reused.
The delegate then paints the wrong values, and the UI is inconsistent with the data model.
This behavior can be observed in this commit when using arrow keys from a confused edit delegate.
Notice the cursor jumping to the sibling of the original spot rendered by the delegate, not a sibling of the current cell.