refactor(oc): revert to TextEdit
I have given up on using a TableView.
Selecting the next cell to edit consistently breaks the UI.

The UI painting issues do not exist in a TextEdit, but there are some downsides.
There is no built-in validator class, so I've resorted to a C++ helper in Keys.onPressed.

Insert/replace modes work, selection & copy works.
The only major feature from #393 that remains unimplemented is auto-formatting.

For future reference, it appears that emitting events inside an item delegate can break the attached delegate.
While I had added a C++ key emitting class, the keypresses happen so close together that the delegate does not fix internal state.
I could have attempted to fix this with an arbitrary QTimer, but this could interfere with user keypresses.