feat: add "x" to close project tabs
My Qt.binding function was never called for tab bar visibility.
The easiest fix required exposing rowCount to QML as a bindable property.

An "X" has been inserted into each normal project tab, allowing for it to be closed independently.
There is still no functionality to trigger save-on-close.
I do not want "+" to be current, but the model deletion logic seems to prevent this on my behalf.