feat(proj): add delegate property to project
The variant contains a delegate which is capable of rendering the current project's edit/debug modes.
Needs to be a variant in C++ so that we get automatic updates on things binding to it.
To attach a property dynamically from QML we need a QtObject, which is incompatible with my data model.