feat(proj): persist object code when switching projects
I had some issues with this feature:
- Directly using onEditingFinished{} would cause an empty string to be
  persisted to the project on a switch.
- Failure to unset sourceComponent would allow one project tab to affect another.

The committed source pattern seems to solve these issues with the minimum amount of code.
An alternative option was a temp project variable in the loader, which was updated in onCurrentProjectChanged.
The source component referred to this delayed variable instead of currentProject.

Partially implements #395 at the ISA3 level.