feat(proj): switch to right-sibling on tab close
This mirrors the behavior of a chrome.

When closing a tab, we need change the currentProject even when the currentIndex does not change.
I am moving functionality to change the current project to the top-level window, for ease of re-use.
The tab bar now calls this function on the main window.