feat(mem): add arbitrary "scroll to:" controls
A text field is linked to the current row of the table.
SP / PC / TextEdit move the view to the indicated row.

We probably want to persist the current address inside of the project so that we don't go back to 0 on tab changes.
PC/SP accessors will need to be provided by the project.

Closes #389.