feat: add (material) icons to mirror Pep9
The included icons were present in Pep9, but I do not know if we need them here.
As they are all vector graphics, I will not feel bad if we have to edit / delete some of these.

Mention material  as a dependency, even though it is not technically required.
Update .gitignore to include `build/` subdirectory.