feat: add main menu for Pep/10 ISA projects
There are some placeholder values for things like recent files and closing projects.

It is not trivial to hide a menu item; it must be removed to be hidden.
With future environments, some menus may have different contents.
Therefore, many of these actions may need to be replaced by Instantiators & signals.