refactor: merge GUI commands into non-singleton
Actions will need to trigger on a specific target, which will not be known at application load time.
I do not feel like storing the current project in a global variable.
Therefore, I need an Actions instance, and it will be less typing to unify the actions into one file.