refactor(sim)!: remove realtime trace analysis
The more I tried to work around this feature, the less useful it seems.

The original goal is that I could write a sink that processes traces and emits breakpoints.
This logic belongs inside the CPU or in the memory devices themselves.
An additional goal was to allow changes in state to change simulator settings.
A better approach would be some kind of signal system, where a signal handler is checked after each sim tick.
Each "thing" that can emit signals would be connected to that handler.
Sample signals:
- breakpoints
- buffer surpassing high watermark

BREAKING CHANGE: unused mode parameter of Buffer::registerSink has been removed.