feat(proj): construct ASMB system with Pep/10 OS
Add parameter to ISA project that prevents it from instantiating a system.
This allows subclasses to implement their own system, but defer to the base class's register/flag bindings.

Assembling the program seems to work, but simulation explodes.
Load times are terrible, so I may disable syntax highlighting for now.