fix(text): fix janky scrolling on Scintilla
It turns out the text editor was being repositioned inside of the viewport on scroll.
That is, a scroll view was created that was as big as the document's logical size.
As you scrolled, the editor would be repositioned within that area.
This has been replaced with some magic scroll bars.

Horizontal scrolling is still not implemented, nor are key strokes.