fix: prevent `wasm-ld` error related to `Pep10_ISA::updateGUI`
The following error only occurs in WASM, and never other builds.
project/pep10.hpp defines Pep10_ISA::updateGUI as a signal with 1-arity.
The MOC correctly generates the implementation for this function.
However pep10.cpp.o ends up with a 0-arity, which I managed to trace to the connect calls.

I discovered that switching to the runtime lookup via SIGNAL/SLOT avoids weird compiler behavior in this case.
As a side effect, it seems to also have fixed single-step debugging not rendering updates in the web version!