fix: improve tooltip for previous memory values
Add an API to separately ask for previous memory values.
I'll have to modify my address sink to generate the necessary info, and this looks hard.
I'll need to create some kind of interval map instead of an interval set.

Hide "previous" entries if there is none recorded.
Progress on #588.