chore(prefs): remove italics, add bold to directive style
This gives it the same appearance as mnemonics.
Per conversation with Stan on 2024-09-14:
We want to differentiate macros from both mnemonics and dot commands.
Pep/9 split dot commands from mnemonics.
It would probably be visual clutter to have italics / underline / bold for the three groups.
So, we are choosing to collapse directives and mnemonics, leaving macros in italics.

Closes #601.