feat: create Pep/9 assembler as duplicate of Pep/10
The "right" thing to do would probably be to add a level of abstraction.
However, I intend to replace the existing assembler with a self-hosting one.
Therefore I view all of this code as temporary.
It better to have it working that have the perfect implementation.

Note that the assembler does not produce useful output at this time.
It just doesn't crash for either Pep/9 or Pep/10