feat(asm.p9): get all Pep/9 figures to assemble
The ELF outputs look good, and there are no errors!
There are magic memory address constants in the `obj` library.
To remove these constants, I would need to add a helper to find the last address before a burn.
This will take some time, and the constants are good enough for now.