feat!: remove Pep/10's loader and `diskIn` port
As per my conversation with Stan conversation 2024-11-09, we want to eliminate the loader from Pep/10's OS.
This toy loader is not of much pedagogical value.
Pep/10 includes a dispatcher which sets up the stack for main, including checking the return value, which is far more impactful than the (infrequently run) loader.

This change will remove ~60 LoC from the OS (about a 10% reduction in size).
Consequently, we will also remove the `diskIn` port, as it was only used for the loader.
From now on, the loader will exclusively be on the C++ of the simulator.

Closes #445, #666
Also resolves: https://github.com/StanWarford/pep9suite/issues/120

BREAKING CHANGE: programs depending on Pep/10s loader or `diskIn` must be refactored.