fix(settings): require mono font for some `PaletteItem`s
This must be validated on a change of parent, on setFont, and on load from JSON.
Otherwise, a non-mono font could make it into our application and break spacing for everything.

PaletteItems are now required to know their category to enable this features.

Also modify font selector to respect request of mono fonts.
On load from JSON, must reset field if it it not explicitly set.
Otherwise, data might propagate between theme changes.

Closes #695.