feat(pas): add FORTH word for decimal input
Ripped from the Pep/10 OS with some modifications:
- Instead of reading from charIn, read from successive characters in _BUF
- Don't mess with NZVC flags, we aren't a system call
- total will be communicated as a return value
- add a new "success" parameter, which will only be false if parsing fails
- don't crash CPU if parsing fails

Future modification would add a single function parameter which would be a pointer to an input buffer.
This would allow the function to be reused on arbitrary strings, not just those in _BUF.

Also, some flags flag isOvfl are useless without NZVC stuff.