feat(pas): allow debugger to step over RETs used for CALL ,s
We don't have a CALL,s but we can push an address onto the stack and return to it.
This is effectively a CALL, minus difficulties in allocating arguments.

However, the simulator does not understand that some RETs effectively increase the stack depth.
I've added a comment tag ;@call, which indicates that a RET is being used to synthesize a CALL.

An assembler operator has been added to gather these abused RETs into set of addresses.
The Pep/10 ISA3 simulator has been extended to use that set to increment or decrement the stack depth accordingly.