refactor(pas): improve decimal read subroutine
It now works in arbitrary bases from 2-36.
I've added a (i16,i16)->i32 multiply function, which is used by the new inti routine to shift properly.
That multiply still seems to have signedness issues, but I will debug them in the future.

I've refactored the control flow of the inti subroutine to be less structured.
It less closely follows the C implementation, but has fewer bytes of object code.

I've also exposed a `*` word to FORTH which uses the new multiplication function.