feat(cpu): add data type selection to CPU registers
A/X/PC/SP can now be viewed as un/signed decimal/ASCII text/binary.
This makes it easier to visualize some of the bit manipulation algorithms I am working with.

I've added a RowFilterModel which selects a single row from an existing model.
Combine this with a TransposeProxyModel, and you can select a single row and make its columns into the content of a QML pane.
I went with this approach because propagating updates was extremely difficult with manually computed model indices.