refactor(proj): break binding between project text and editors
Editors should be the source of truth, only synced to the project when they need to be persisted or operated on.

This should prevent jank caused when clicking out of the editor after making changed.
Before, you always scrolled to the top.
Now, you should remain in place.
This jank was caused by round-tripping the text from the editor to the project back to the editor.