fix(bi): re-enable "Copy to Project"
Usually, data flows unidirectionally from the editor to the project.
However, the help system is allowed to replace the "guts" of a project.
Using the overwriteEditors signal, we can have a one-off sync from the project to the editor, after which unidirectional dataflow resumes.

Without this modification, "copy to project" is a useless button.