feat(welcome): split architecture and project selection
There are now two levels, one for the architecture and one for the project.
The default is Pep/10, and it is sticking between app launches.

Some of the UI/UX is hindered by presence of placeholder data, but I will iterate after checking this in.
There are visual hacks to work around the unfiltered model that we currently have.

There are still too many abbreviations, too few pictures, and too much text.

But, it's still progress on #765.