build: fix enitlements for Mac OS x86 codesigning
While we are not electron, the following issues points out some steps
to fix notarization:
https://github.com/electron-userland/electron-builder/pull/3926

On Stan's machine Pepp passes both notarization and codesigning on his
machine.
This means there is a runtime issue with the app, likely related to
the JS JIT stuff.

Maybe closes #766?
Will have to wait for CI passing and cutting a release to be sure.