feat(symbol): enable cell selection in table view
Had to switch to a TableView to enable a rectangular selection.
Switched from a ListModel to a TableModel as a requirement of using a TableView.
TableView is responsible for calculating column widths & the number of columns that can fit on the display.
It must inform the model of this column count because TableViews don't supporting "wrapping" like GridViews do.

Inlined SymbolCell to remove unnecessary nesting.

Added a helper method to the symbol model to facilitate rectangular selections.
This really belongs on a "Utils" class, but I don't want to make another hpp/cpp pair for this.

Progress on #776.