feat: use fonts to distinguish macro expansions
Improve the lexer to count the number of start / end macro comments.
If the counter is positive, we set a bit flag in our style indicating that we are in a macro.
Styling inside and out of a macro is identical except that that macros use a mechanical-looking font.

We might not keep this feature, but I thought it was worth a shot giving macro expansions a separate "voice".