feat(bi): add tokenizer
I'm still not happy with how "pythonic" the tokenizer is.

I feel like the tokenizer should be an iterator/generator implementing next() rather than having an explicit next_token().
Also, I feel I need the ability to peek the stream, and some refactoring to indicate we've reached the end of iteration.