feat(bi): allow loading builtin registry from arbitrary folders
This allows programmers to select from where figures should be loaded.
It will be used to implement the "hot reloading" from #790.

I've modified the default behavior of the registry CTOR to not crash on invalid figures or books.
Instead, errors are emitted to stderr, and the broken book / figure / macro is ignored.
Permitting these errors instead of crashing should provide Stan more flexibility in authoring figures.