fix(bi): prevent console spam on loading external figures
I was trying to treat the special files `.` and `..` as input/output pairs.
The OS was not pleased with my attempt to read them as text files, and caused many warning to appear on my console.

This did not happen previously with the QRC system, because those special files did not exist.