feat(bi): enable reloading of figures at app runtime
This isn't the prettiest implementation.
I don't like the hardcoded index variables.

I tried an approach with find and iterators, but I was getting segfaults.
Those segfaults were likely related to use-after-free in data, but I never went back to check.

In the future, if I have more things to reload, I will need a more robust system.

Progress on #790.