refactor: allow '\0' as null terminator in Pep/10
This is isomorphic to C/C++, where \0 create a null terminator in a string.
In C, this is because \0 is a valid octal escape sequence.
For Pepp, we will treat \0 as a special case and ignore other octal escape sequences.

Hex escape codes as still (invisibly) available, but \0 is the canonical way to create a null terminator.
Program sources & listings will respect the form of null terminator that you create.

Pep/9 now accepts this form of null terminator as well.
This will remain an undocumented feature.