feat(sim): begin adding packrat-style memoization
To prevent combinatorial explosion of function calls, we should remember if we've already attempted a rule given an input position.

This commit doesn't implement the caching algo, but it does start to record the info I will need to make the cache work.

When this is finished, each rule should be visited no more than once for a given input position.