feat(sim): cache interim parsing results in expression evaluator
Add caching to prevent checking the same token index under the same rule more than once.
Previously, the computational complexity was roughly O(3^N), where N is the number of rules.
With this parser, we should be polynomial time at the cost of polynomial space.

I've formulated my cache such that it is only a performance optimization.
This is unlike the tables in packrat parsing which are required for correct execution.

I still think there's room to improve my grammar to reduce complexity, but this is a good start.

This addresses performance concerns from f253594a9d2eedfca420b8816928b159210bf432.
I still need to resolve the issues of `a == b == c`.