fix(sim): allow expressions with multiple same-precedence operators
I had a bug with my recursion where a == b == c did not work because the RHS recursion always goes down a level.
My grammar fix of recursing to `expr` also turned out to be wrong and inverted my operator precedence.

With some debugging, I figured out the RHS rule needs to be the same precedence.
This lead to passing tests.