refactor(sim): prevent mutating AST members depend on in <=>
As part of AST's interaction with the expression cache, AST nodes need to been immutable after insertion into the cache.
More specifically, fields which we sort on via <=>.

Prior to this commit, this has only be enforced by the convention "don't do that".
Now, I've marked fields on which we sort as const.
Future modifications to each node and <=> must continue to mark sorting key material as const.

Going the const-member route lets me skip a bunch of accessor methods.