feat(sim): add sized integer literal suffixes to parser
Typing information for bits has been moved into its own header so that the lexer does not depend on evaluator headers.

Suffixes are {_i8, _u8, _i16, _u16, _i32, _u32}.
If not directly attached to an integer literal it will instead be parsed as an identifier.
Suffixes are generated as a separate token from the attached numeric literal for ease of str=>int conversions in lexer.

Does not replace the need for type casting, but it is at least a start towards more interesting expressions.