fix(sim): recompute watch expressions on simulation start
At the start of the simulation, all rows in the watch expression pane should be re-evaluated and made clean.
This closes #805.

Also fix a bug where a value changed multiple times in a row would only update if it stopped changing.
Fixed by replacing a ^ b with a | (a ^ b).