feat(sim): update breakpoint viewer on add / remove
Change fundamental data structure of BPs from set to vector.
It will be easier to iterate / assign rows to items in vector than set.
The vector is required to be sorted (like the set), so lookups are still lg(n), at the cost of nlg(n) inserts/deletes.

All values except address are wrong.
however, it reacts to breakpoint inserts / deletes.