refactor(sim): merge OS+User symbol table models
The purpose of this refactor is to make it easier to expose the static value of symbols via the debugger.

So, I want a single model which represents both user and OS symbols.
Applying a filter will help make the presentation manageable / match todays presentation.

But holy mother of proxy models batman.
For filtering to work best, my data needs to be a 1D table (list).
Because filters are per row / column, but not per index.

Then, I can proxy that filtered data into a 2D table, where I define all common model index operations.
This can't be combined with the filter, because in reimplementing index ops, calls to `filterAcceptsRow` were removed.

Multiple hours later, we have this triply nested model.
And if all went well, everything looks the same.