feat(sim): extract editable watch expressions to own class
I've discovered that it will be exceedingly hard to reuse WatchExpressionEditor for conditional BP expressions.
However, the underlying data structure, an editable watch expression, is easy to re-use between both.

I've modified the important algorithmic bits to be free functions operating on iterables of EditableWatchExpression.

This enabled me to reuse most of my editing logic between the watch expression and breakpoint viewer panes.