feat(sim): add placeholders for trace tag extraction
Need to write the visitor to extract trace tags from comments.
Trace tags need to be evaluated to see if they are missing a command and infer the command if absent.
Then, need to assign those trace tags to addresses / global scope.
Lastly, need to serialized to the elf file.

Then I can start working on the simulator side of things.