build: cache changelog DB between builds
Do not replace existing changelog DB if there have been no updates.
Replacing that DB causes Qt's resource compiler to rebuild the QRC file, which leads to re-linking unnecessary targets.
My hash over the database is conservative -- it may falsely state that two DBs are different when they are the same.
This will cause a spurious rebuild, but is not a correctness issue.

This will prevent long link times on hitting "run" in the IDE when no changes were made.