refactor(welcome): redo layout & design of welcome screen
Text is now displayed as a tooltip on hover or button press.
Buttons now contain 3 lines: 1) architecture, 2) abstraction, and 3) additional features, if enabled.
Icon and button have mostly been merged.
"cards" are now arranged in a 3 column by N row setup.

I may need to limit the number of columns or make the whole thing scrollable if it doesn't "fit" on the display.

Progress on #847.