refactor: use docking widgets for ISA3
All "parts" of the ISA3 project are now DockingWidgets, allowing users to resize at their leisure.

Required refactoring the CPU pane to not use a Layout, otherwise we have no clue how tall the pane is.
This has re-introduced alignment and clipping bugs, which I will fix later.

I want to test this code on Windows/WASM to make sure the changes otherwise behave as expected.