fix: preserve changes to component visibility for a given mode
Assume a user marks the "Object Code" pane as visible in the Editor mode.
If they switch to the debugger then back to the editor, prior to this commit, that pane would have been hidden.
I've added some fields to each dock widgets to help preserve visibility information.
I could extend this approach to remember the layout/size of widgets in each mode.