refactor(welcome): layout according to diagram in #856.
TL;DR, we don't like the 2-level architecture/project selection.
We want project selection to be flat.

To start off, we put "Computer System <> Edition" as the title in the welcome bar.
Then, we insert buttons for 6,5,4 in the <> to let you cycle between editions.

I needed to refactor my settings to allow storing a default edition of the book, not just an architecture.

I added a negative top margin so that welcome/help appear in place of the toolbar.
Closes #856.