build: enable iOS builds
Stop building test executable unconditionally, otherwise the iOS version will not compile.
Will not be tested in CI for now, as I do not plan to deploy to the app store yet.
Focus is terrible, menus are wacky, and Asmb projects crash on creation.

At least it builds, which is a step in the right direction?