feat(sim): prevent direct evaluation of Terms
Multiple locations will now be evaluating Terms: watch expressions, breakpoints, and the stack trace.
I want a centralized class which helps me figure out if the value has changed since the last time I called `evaluate`.
To that end, Term::evaluate is now protected, and must be accessed via a CachedEvaluator.
If you hold on to a CachedEvaluator, it is trivial to determine if there has been an update to a value.

I've added a `version` field to an ExpressionCache.
Anytime you do a meaningful calculation in evaluate, increment `version`.
The CacheEvaluator can compare its version to the Term's version, validating up-to-dateness.

A new `UseDirtyAlways` caching policy has been added, which always returns the cache if it has been set.

Progress on #801.